BUG FIX: fixed a bug in the spacegroup name reformatting code in the cif2cod and cif_filter scripts.
Also, the two scripts now do not try to reformat a H-M spacegroup
symbol if tht symbol is not found in the spacegroup list.

Spaces are removed from the 'text' field value before printout in the
cif2cod script.

The cif2cod script no longer reports an error if the
_journal_page_last tag is missing from the bibliographic data.


git-svn-id: svn://cod.ibt.lt/cod-tools/trunk@362 04be6746-3802-0410-999d-98508da1e98c