BUG FIX: now the cif_filter at least does not glue a tag and a long (foldable) space-less value, it folds the value before the printout and correctly determines what separator (a space or a newline) should be printed after the tag. This also fixes formally correct but ugly printouts in some cases, were a ;;-fields value was separated from its tag by an extra, unnecessary empty line.
The formatting is not ideal, as shown in the
outputs/cif_filter_036.out test's _publ_section_title tag -- the tg
lost its ;;-signs upon folding since it had no spaces in it. This is
formally correct but not nice. Also, the '#\#' CIF filding method is
not implemented. So far, the script can be used for depositions and
will yield vcif-clean results in most cases, but the policy of CIF
printout and value folding must be reconsidered and maybe the code
radically redesigned.


git-svn-id: svn://cod.ibt.lt/cod-tools/trunk@427 04be6746-3802-0410-999d-98508da1e98c