Adding the space_groups.dp.PL progam from the Atoms 3.0 package. Strangely, the source archive is not called atoms-x.y.z.tgz, but horae-066.tar.gz. I have downloaded it from http://cars9.uchicago.edu/~ravel/software/packages/horae-066.tar.gz recently; as of today, the link is still there.
The file header says:

All rights reserved. This program is free software; you can
redistribute it and/or modify it under the same terms as Perl
itself.

I appreciate greatly this prograemchen since it has a full list of all
ITC space groups, origins, and settings. This saves me *a lot* of work
and time! Thanks, Bruce!


git-svn-id: svn://cod.ibt.lt/cod-tools/trunk@447 04be6746-3802-0410-999d-98508da1e98c