Extracting the spacegroup list variable, %space_groups, verbatim from the space_groups.db.PL program, and converting it into a Perl module Spacegroups.pm. The extraction command was:
perl -ne 'print if /^my %space_groups/../^\s*\);/' space_groups.db.PL

The Perl module Spacegroups.pm with the extracted data was then used
in the mknames.pl Perl script to produce a spacegroup list in the
format used in the SpacegroupNames.pm module. This generated list was
then manually appened to the SpacegroupNames.pm spacegroup array. It
is now ready to be used in other cif-tools Perl scripts.


git-svn-id: svn://cod.ibt.lt/cod-tools/trunk@448 04be6746-3802-0410-999d-98508da1e98c