Changing one seting of the 'c 2' group to what I think is corect one:
For the spacegroup 'c 2' the file Spacegroups/Spacegroups.pm, following 
the space_groups.db.PL from horae-066/Xtal/, specifies 'b 2 1 1' as the 
next-to-the-last elemet of the 'settings' array 
($space_groups{c 2}{settings}[7]). But the short form says it is 'c 2' 
($space_groups{c 2}{short}[7]). Shouldn't the full name for this setting 
be 'c 2 1 1' not the 'b 2 1 1'?


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@461 04be6746-3802-0410-999d-98508da1e98c