Adding full universal Hermann Mauguin designators for rhombohedral groups into the SpacegroupNames.pm module. Chamging the module policy to list the original, unchanged group name in the third column; thus, the spacegroup renaming scripts (such as cif_filter) should leave the original spacegroup names unchanged.
Previously, the therd column would list the full H-M symbols, and the
second column would list the short H-M symbols, and the cif_filter
would expand a spacegroup name into it's full form.

The reason for such policy change is a mess in the nomenclature of the
rhombohedrasl (R*) groups; different authors suggest different
designations for the same setting (e.g. PDB use H3 designator and
CCTBX use R3:H), and it is not clear what setting the 'R 3' designator
actually conveys.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@715 04be6746-3802-0410-999d-98508da1e98c