Adding the simple chemical formula parser from the svn+ssh://narvalas.ibt.lt/home/saulius/svn-repositories/formula_sum repository revision 1.
This is the first working version of the CIF chemical formula parser
in Perl. It is not yet accepting all correct formulas (e.g. it
requires a space in front of parentheses), and accepts (deliberately)
some incorrect formulas (e.g. formulas without spaces, with "glued"
element symbols, like in C2H6O); but in spite of this it is very handy
to check and possibly restor CIF _chemical_formula_sum from various
author supplied formulae.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@738 04be6746-3802-0410-999d-98508da1e98c