trunk/:
Changing behaviour of the CIFTagManage.pm, which afects, most
importantly, the behaviour of the cif_filter script (but also of all
other ecripts that might use this module).

Now the exclude_empty_tags subroutine excludes only tags that are not
in loops. This is done to avoid emissin empty but compulsory loop
tags, since such omission would produce CIFs that fail to validate
against core dictionaries.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@769 04be6746-3802-0410-999d-98508da1e98c