trunk/:
Changing printout of the CIF values in the CIFTagPrint.pm module: now
it enquotes string values starting with a hash ("#") symbol, to
distinguish them from comments in the output CIFs. This changes
behaviour of all scripts using this module, notably of the cif_filter
script, and the new behaviour is arguably more correct.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@772 04be6746-3802-0410-999d-98508da1e98c