trunk/: FUNCTIONALITY FIX: now the CIFParser returns all tags converted into lowercase, not in their original spelling. Such behaviour is needed to process tags with unusual capitalisation correctly. The scripts that want to output mixed-case tags should use tag name canonicalisation from the CIFTagCanonicalNames.pm module; examples of such scripts are cif_filter and cif_validate.
The price to pay is that now when an unknown tag is encountered and 
reported, its name is printed in all lowercase, not in the original 
spelling. I was so far hesitant to store and process original spellings, 
since this would put unnecessary burden on our CIFParser, and is not 
needed by most of the CIF computing programs. If the unprecise report of 
misspelled tags is an issue, however, some mechanism to store original 
tag names will have to be built in into the CIFParser.


git-svn-id: svn://cod.ibt.lt/cod-tools/trunk@825 04be6746-3802-0410-999d-98508da1e98c