trunk/ (saulius@koala.ibt.lt) Merging the C CIF parser speadup code:
svn merge -r2405:2498 branches/saulius-C-CIF-parser-speedup/ trunk/

The conflicts were resolved manually. Currently, the behaviour of the 
parser in the trunk has changed, and it can no longer fix stray text 
fields in the beginning of the file. For this reason, two tests do not 
pass. It remains to be seen how we deal with this.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@2500 04be6746-3802-0410-999d-98508da1e98c