trunk/ (saulius@koala.ibt.lt) Making the dash ('-') between the error message and the explanation text explicitely required.
Also, making spaces after the separating dash obligatory. This change in 
the grammar rejects some error messages as incorrect, but for a good 
reason -- in these messages, the minus sign of a number was previously 
mis-parsed as a message part separator.

minus signs will have to be escaped using XHTML entities; this is 
awkward for negative numbers, but for now I have no better solution 
(another solution would be to replace the separator character '-' by 
somthing that is used only seldom).


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@3571 04be6746-3802-0410-999d-98508da1e98c