trunk/ (saulius@kukutis) Changing the behaviour of the 'cif_filter' program: it now outputs the CIF header only once at the beginning of the whole output data stream instead of placing them in front of each data block.
This is deemed to be a correct behaviour, since:

- the '#\#CIF_2.0' format version header is only effective at the 
beginning of the file;

- the current 'cif_split_primitive' duplicates the first header at the 
beginning of each split file, and outputs all comments, resulting in 
dublicated header information.

This behaviour assumes that the output stream will be of the same syntax 
regardless of how many and what version are the listed input CIFs.


git-svn-id: svn://www.crystallography.net/cod-tools/trunk@9508 04be6746-3802-0410-999d-98508da1e98c